Add files via upload
Data Source and Dataset Information
The dataset is related with the marketing campaigns of a Portuguese banking institution. 
The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or would not be ('no') subscribed.

Objective: The purpose of this project is by using machine learning algorithm we need to find out if the bank deposit term would be subscribed (‘yes’) or not (‘no’) by a client. Term deposit is defined as fixed-term investment that includes the deposit of money into an account at a financial institution. 

Data Observations
The following categories of information are included in the data set: 
•	age (numeric)
•	job: type of job (categorical: “admin”, “unknown”, “unemployed”, “management”, “housemaid”, “entrepreneur”, “student”, “blue-collar”, “self-employed”, “retired”, “technician”, “services”)
•	marital: marital status (categorical: “married”, “divorced”, “single”; note: “divorced” means divorced or widowed)
•	education (categorical: “unknown”, “secondary”, “primary”, “tertiary”)
•	default: has credit in default? (binary: “yes”, “no”)
•	balance: average yearly balance, in euros (numeric)
•	housing: has housing loan? (binary: “yes”, “no”)
•	loan: has personal loan? (binary: “yes”, “no”)
•	contact: contact communication type (categorical: “unknown”, “telephone”, “cellular”)
•	day: last contact day of the month (numeric)
•	month: last contact month of year (categorical: “jan”, “feb”, “mar”, …, “nov”, “dec”)
•	duration: last contact duration, in seconds (numeric) # other attributes:
•	campaign: number of contacts performed during this campaign and for this client (numeric, includes last contact)
•	pdays: number of days that passed by after the client was last contacted from a previous campaign (numeric, -1 means client was not previously contacted)
•	previous: number of contacts performed before this campaign and for this client (numeric)
•	poutcome: outcome of the previous marketing campaign (categorical: “unknown”, “other”, “failure”, “success”)
Output variable: y - has the client subscribed a term deposit? (binary: “yes”, “no”)